Began Flight class implementation
Began the implementation of the Flight class. Default and direct-value
constructors are fully implemented. The string constructor is a work in
progress and not at all functional currently.

The departure and destination cities are stored as a string. The cost is
stored as integer. The departure and destination times are stored as
integers. The current plan is to implement time as a number of minutes
since midnight.