Time class and flight schedule graph prep
Created the Time class which will be used to store all times for the
Flight class. Implemented +, -, ==, and << operators.

Converted the departureTime and destinationTime data members of the
Flight class to type Time. Converted the cost data member to type float.
Changed existing constructors and created a new constructor to reflect
this change. Changed the "Time()" function to "timeStringToInt()" so as
to prevent confusion with the Time class.

Changed the Trip class to contain a vector of Flights called itinerary.
Added a function addFlight(Flight) which adds a specified flight to the
itinerary.

Moved the readFile function to Program3 along with the checkDate,
checkTime, and checkCity functions. the readFile function may be moved
depending on the implementation of the flight schedule structure.